pyproject.toml
[project]
name = "omni-home"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-environ>=0.11.2",
    "django>=5.1.4",
    "gunicorn>=23.0.0",
    "dj-database-url>=2.3.0",
    "psycopg2>=2.9.10",
    "whitenoise>=6.8.2",
    "django-crispy-forms>=2.3",
    "crispy-tailwind>=1.0.3",
]

[dependency-groups]
dev = [
    "pytest-django>=4.9.0",
    "ruff>=0.8.2",
]
